+++ cmd/command_darwin.go
@@ -1,0 +1,7 @@
+package cmd
+
+import "github.com/robocorp/rcc/pretty"
+
+func osSpecificHolotreeSharing(enable bool) {
+	pretty.Warning("Good to go. Nothing to do on Mac OS.")
+}