+++ pathlib/lock_windows.go
@@ -54,7 +54,7 @@ func Locker(filename string, trycount int) (Releaser, error) {
 			time.Sleep(40 * time.Millisecond)
 			continue
 		}
-		_, err = MakeSharedFile(filename)
+		_, err = shared.MakeSharedFile(filename)
 		if err != nil {
 			return nil, err
 		}